Replaced ASSERT exception with call to exit()
Using exit function in place of an exception since the only action that
can be taken after an exception error is to clean up and exit. In addition,
each assertion (injected by ASSERT) is not a part of the product code. Thus,
introducing try/catch blocks to catch assertion errors does not make sense.