BUG FIX: Fixed Vector::zero function
The purpose of the Vector::zero function is to set all elements of the
vector to zero. However, for some silly reason, it did this by way
of inplace multiplication with a zero element (data[i] *= zero).
Provided that the internal data array was properly initialized to some
sensible numerical value, this is fine. However, if the default values
correspond to 'inf' or 'nan', the multiplication will yeild
unpredictable results (not to mention the inherent inefficiency).

The implementation now properly uses the broadcast assignment operator
build-in to the underlying array container. In principle, the array
container should implement the 'best' strategy for executing this kind
of assignment.