FIX: Corrected the implementation of the sign function
Note that since 'sign' was previously defined (though incorrectly),
the name was switched to 'sgn'. Thus, if any other working code makes
use of the former 'sign' function a compilation error will result.
Any users of the old 'sign' function which require the "incorrect"
defition should implement a separate function which does not use
the 'sign' or 'sgn' name. The 'sign' function has a former mathematical
definition. The NumLib implementation of this function should be
consistent with its formal definition.

Also note that GateauxFD made use of the old 'sign' function, but
required the inconsistent definition. A small in-place modification
of GateauxFD was made which should not modify its behavior.