Modified initialization of particles in OptimizerPSO
The particles used in PSO are now given explicit initial conditions by
the optimizer. This allows one particle to be placed at the user
provided estimate for the optimum. It also allows the possibility of
using a uniformly spaced distribution of particles instead of random;
although, at the present time use of a uniform random distribution
would appear be sufficient, if not neccessary.