Removed incomplete unit testing suite
This was done to prepare the git tree for the new branching
scheme which facilitates the incorporation of version numbering
into the Git repo via tags on the master branch. The Boost unit
tests deleted here are still available in the working tree but
will remain as untracked until they've been completed (i.e. a
baseline implementation has been completed).