Changed solve template function to regular function for SquareMatrix<Real>
Since the only the template specialization for Real types is implemented
the generic template function is quite useless.

In future, the template solve function can be reintroduced in the future once
a generic LU decomposition algorithm has been implemented for generic scalar types.
Alternatively, once can simply overload the solve function for each scalar type in
which there is a LAPACK algorithm that can handle it (e.g. complex numbers).