Added stream manipulators (based on old io.h) file
This is really just a cleanup of the old io.h header. In practice,
the boost::format library is recommended over the manipulator(s)
defined here. However, there is some other code already in use which
uses the fmtReal from the old io.h header. So, I've decided to keep
that manipulator availible for some time to come.