Changed ArrayContainer to use () for subscripting instead of []
This was done to be consistent with the array and linalg libraries
and to be able to resuse the interpolation schemes in the reconstruction
library. It break's STL conventsions, but in NumLib the () operator has too
much inertia and trying to break the trend now is causing too many conflics.

The reason for choose () over [] was so that a multidimensional array A
could be accesses via A(i,j) instead of A[i][j]; the latter of which was
thought to be less efficient when contiguous storage is used. The reason
for wanting contiguous storage was to be compatible with Fortran libraries.

The issue of using () versus [] for array subscripting can be revisited in
the next major release.