BUG FIX: Corrected computation of uTv_sign in GateauxFD operator
This was actually a bug introduced when the numlib::sign was replaced
with numlib::sgn. Originally, numlib::sign was implemented as the function
that returns +1 on [0,+infty) and -1 otherwise--which is exactly the behavior
we want for GateauxFD. However, by definition, the "sign" or "sgn" function should
return 0 when the argument is 0. To make a long story short, another application
required the correct sign (sgn) function implementation, so numlib::sign was removed
and numlib::sgn was implemented per the correct mathematical definition; numlib::sign
was removed to avoid confusion. When this change was made, GateauxFD was modified to
compute "sign" as the value of the argument over its absolute value; this is a bug since
it results in 0 when the argument is 0 (we want +1 when the argument is 0).